requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-CORS
Flask-Login
PyMySQL
cryptography
python-dotenv
sqlalchemy